+++ test/integration/src/pages/http/CookieClient/CookieClient.js
@@ -2,21 +2,18 @@ import React from 'react';
 import Page from 'soya/lib/page/Page';
 import RenderResult from 'soya/lib/page/RenderResult';
 import register from 'soya/lib/client/Register';
-import { DevTools, DebugPanel, LogMonitor } from 'redux-devtools/lib/react';
 
 class Component extends React.Component {
   render() {
     return <div>
-      <DebugPanel top right bottom>
-        <DevTools store={this.props.store} monitor={LogMonitor} />
-      </DebugPanel>
+
     </div>;
   }
 }
 
-class AddCookie extends Page {
+class CookieClient extends Page {
   static get pageName() {
-    return 'AddCookie';
+    return 'CookieClient';
   }
 
   render(httpRequest, routeArgs, callback) {
@@ -24,5 +21,5 @@ class AddCookie extends Page {
   }
 }
 
-register(AddCookie);
-export default AddCookie;
+register(CookieClient);
+export default CookieClient;