+++ test/integration/src/pages/redux/Hydration/Hydration.js
@@ -76,7 +76,7 @@ class Hydration extends Page {
   }
 
   createStore(initialState) {
-    var reduxStore = new ReduxStore(Promise, initialState, this.config, this.cookieReader);
+    var reduxStore = new ReduxStore(Promise, initialState, this.config, this.cookieJar);
     return reduxStore;
   }