+++ test/integration/src/pages/redux/SegmentCookie/SegmentCookie.js
@@ -26,7 +26,7 @@ class SegmentCookie extends Page {
   }
 
   createStore(initialState) {
-    var reduxStore = new ReduxStore(Promise, initialState, this.config, this.cookieReader);
+    var reduxStore = new ReduxStore(Promise, initialState, this.config, this.cookieJar);
     return reduxStore;
   }