+++ test/integration/src/pages/redux/ServerSegmentDependencies/ServerSegmentDependencies.js
@@ -64,7 +64,7 @@ class ServerSegmentDependencies extends Page {
   }
 
   createStore(initialState) {
-    var reduxStore = new ReduxStore(Promise, initialState, this.config, this.cookieReader);
+    var reduxStore = new ReduxStore(Promise, initialState, this.config, this.cookieJar);
     return reduxStore;
   }